(PC-19590) refactor(errors): make Errors following the code structure
