--- src/features/errors/hocs/index.ts
@@ -1,1 +1,0 @@
-export * from './withAsyncErrorBoundary'
