--- src/features/errors/index.ts
@@ -1,2 +1,0 @@
-export { AsyncErrorBoundary } from './pages/AsyncErrorBoundary'
-export { withAsyncErrorBoundary } from './hocs/withAsyncErrorBoundary'
