+++ src/features/errors/pages/AsyncErrorBoundary.native.test.tsx
@@ -5,7 +5,7 @@ import { mockGoBack } from 'features/navigation/__mocks__/useGoBack'
 import { AsyncError } from 'libs/monitoring'
 import { render, fireEvent } from 'tests/utils'
 
-import { AsyncErrorBoundary } from '../AsyncErrorBoundary'
+import { AsyncErrorBoundary } from './AsyncErrorBoundary'
 
 jest.mock('libs/monitoring/services')
