+++ src/features/navigation/RootNavigator/routes.ts
@@ -29,7 +29,7 @@ import { NavigationProfile } from 'features/cheatcodes/pages/NavigationProfile'
 import { NavigationShareApp } from 'features/cheatcodes/pages/NavigationShareApp/NavigationShareApp'
 import { ThematicHomeHeaderCheatcode } from 'features/cheatcodes/pages/ThematicHomeHeaderCheatcode/ThematicHomeHeaderCheatcode'
 import { EighteenBirthday } from 'features/eighteenBirthday/pages/EighteenBirthday'
-import { withAsyncErrorBoundary } from 'features/errors'
+import { withAsyncErrorBoundary } from 'features/errors/hocs/withAsyncErrorBoundary'
 import { BannedCountryError } from 'features/errors/pages/BannedCountryError'
 import { FavoritesSorts } from 'features/favorites/pages/FavoritesSorts'
 import { CulturalSurvey } from 'features/firstLogin/CulturalSurvey'
